pyproject.toml
[tool.poetry]
name = "assistantgpt"
version = "0.1.0"
description = "Your personal assistant, powered by chatGPT"
authors = ["astelmach01 <astelmach20@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rich-click = "^1.6.1"
python-dotenv = "^1.0.0"
langchain = "^0.0.240"
openai = "^0.27.8"
pydantic = "<2.0"
flake8 = "^6.0.0"
pytest = "^7.4.0"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
